+++ backends/candle/src/layers/mod.rs
@@ -2,6 +2,7 @@
 mod cublaslt;
 mod layer_norm;
 mod linear;
+mod radix_mlp;
 #[allow(dead_code, unused)]
 mod rms_norm;
 mod rotary;
@@ -9,6 +10,8 @@ mod rotary;
 pub use cublaslt::get_cublas_lt_wrapper;
 pub use layer_norm::{LayerNorm, LayerNormNoBias};
 pub use linear::{HiddenAct, Linear};
+#[allow(unused_imports)]
+pub use radix_mlp::CompactUnfoldTensors;
 #[allow(unused_imports)]
 pub use rms_norm::RMSNorm;
 pub use rotary::{apply_rotary, get_cos_sin, get_inv_freqs, RopeScaling};
